+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## v4.94.0 Features
 
 - `[Homepage]` Added row height style classes for widgets. ([#8211](https://github.com/infor-design/enterprise/issues/8211))
+- `[Lookup]` Added placeholder setting in toolbar. ([#8416](https://github.com/infor-design/enterprise/issues/8416))
 
 ## v4.94.0 Fixes